added more integration tests

+++ Store.API.IntegrationTests/TestWebApplicationFactory.cs
@@ -1,19 +1,18 @@
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 using StoreAPI.Persistence;
 using System;
 using System.Linq;
+using System.Threading.Tasks;
 
 namespace Store.API.IntegrationTests
 {
-    public class TestWebApplicationFactory<TStartup> : WebApplicationFactory<TStartup> where TStartup : class//, IDisposable
+    public class TestWebApplicationFactory<TStartup> : WebApplicationFactory<TStartup> where TStartup : class
     {
-        public object Lock { get; set; } = new object();
-        public object DbLock { get; set; } = new object();
-        private static bool _databaseInitialized;
-
         public string DatabaseName { get; }
 
         public TestWebApplicationFactory()
@@ -28,14 +27,18 @@ namespace Store.API.IntegrationTests
 
         public void ConfigureServices(IServiceCollection services)
         {
-            var descriptor = services.SingleOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<StoreDbContext>));
-            services.Remove(descriptor);
+            var dbContextDescriptor = services.SingleOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<StoreDbContext>));
+            var configJsonOptions = services.SingleOrDefault(sd => sd.ServiceType == typeof(IConfigureOptions<JsonOptions>));
 
-            var sp = services.BuildServiceProvider();
+            services.Remove(configJsonOptions);
+            services.Remove(dbContextDescriptor);
 
             services.AddScoped<StoreDbContext>(s => CreateContext());
 
-            EnsureDeleteAndCreateDatabase();
+            var sp = services.BuildServiceProvider();
+            var context = sp.GetRequiredService<StoreDbContext>();
+
+            context.Database.EnsureCreated();
         }
 
         public StoreDbContext CreateContext()
@@ -51,31 +54,22 @@ namespace Store.API.IntegrationTests
                 .Options);
         }
 
-        public void EnsureDeleteAndCreateDatabase()
-        {
-            //lock (DbLock)
-            {
-                using var context = CreateContext();
-                _databaseInitialized = false;
-                context.Database.EnsureDeleted();
-                context.Database.EnsureCreated();
-
-                _databaseInitialized = true;
-            }
-        }
-
-        public void Cleanup()
+        public async Task Cleanup()
         {
             using var context = CreateContext();
-            context.Database.EnsureDeleted();
-            context.Database.EnsureCreated();
+            await context.Database.EnsureDeletedAsync();
+            await context.Database.EnsureCreatedAsync();
         }
+
 
         protected override void Dispose(bool disposing)
         {
-            using var context = CreateContext();
-            context.Database.EnsureDeleted();
-            base.Dispose(disposing);
+            if (disposing)
+            {
+                using var context = CreateContext();
+                context.Database.EnsureDeleted();
+                base.Dispose(disposing);
+            }
         }
     }
 }
